feat: Jdbc를 JPA로 마이그레이션

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -48,8 +48,7 @@ public class ReservationService {
                 .orElseThrow(() -> new NotFoundException("존재하지 않는 예약 시간입니다."));
         final Theme theme = themeRepository.findById(request.themeId())
                 .orElseThrow(() -> new NotFoundException("존재하지 않는 테마입니다."));
-        if (reservationRepository.existsByDateAndTimeAndTheme(request.date(), reservationTime.getStartAt(),
-                theme.getId())) {
+        if (reservationRepository.existsByDateAndTimeAndTheme(request.date(), reservationTime, theme)) {
             throw new ReservationException("해당 시간은 이미 예약되어있습니다.");
         }
 
@@ -66,8 +65,7 @@ public class ReservationService {
                 .orElseThrow(() -> new NotFoundException("존재하지 않는 테마입니다."));
         final Member member = memberRepository.findById(request.memberId())
                 .orElseThrow(() -> new NotFoundException("존재하지 않는 멤버입니다."));
-        if (reservationRepository.existsByDateAndTimeAndTheme(request.date(), reservationTime.getStartAt(),
-                theme.getId())) {
+        if (reservationRepository.existsByDateAndTimeAndTheme(request.date(), reservationTime, theme)) {
             throw new ReservationException("해당 시간은 이미 예약되어있습니다.");
         }
         final Reservation reservation = new Reservation(request.date(), reservationTime, theme, member);
@@ -76,9 +74,13 @@ public class ReservationService {
     }
 
     public void deleteReservation(final Long id) {
-        final int deletedCount = reservationRepository.deleteById(id);
-        if (deletedCount == 0) {
-            throw new NotFoundException("존재하지 않는 예약입니다. id=" + id);
+        /*final int deletedCount =*/
+        if (!reservationRepository.existsById(id)) {
+            throw new NotFoundException("예약이 존재하지 않습니다.");
         }
+        reservationRepository.deleteById(id);
+//        if (deletedCount == 0) {
+//            throw new NotFoundException("존재하지 않는 예약입니다. id=" + id);
+//        }
     }
 }
